requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
black==21.6b0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
Django==3.2.5
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2021.1
regex==2021.7.6
schedule==1.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
wincertstore==0.2
